Move assets/icons to dedicated folder (#4850)
